+++ lib/Screens/Menu/placedMealMenuOrders.dart
@@ -1,60 +1,76 @@
-import 'dart:convert' as convert;
-
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:date_format/date_format.dart';
-import 'package:diet_delight/Models/foodItemModel.dart';
 import 'package:diet_delight/Models/mealModel.dart';
 import 'package:diet_delight/Models/mealPurchaseModel.dart';
 import 'package:diet_delight/Models/menuCategoryModel.dart';
 import 'package:diet_delight/Models/menuModel.dart';
 import 'package:diet_delight/Models/menuOrdersModel.dart';
+import 'package:diet_delight/Screens/Menu/placeMealMenuOrders.dart';
 import 'package:diet_delight/konstants.dart';
 import 'package:diet_delight/services/apiCalls.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_dropdown/flutter_dropdown.dart';
+import 'package:syncfusion_flutter_datepicker/datepicker.dart';
 
-class EditMealMenu extends StatefulWidget {
+class PlacedMealMenuOrders extends StatefulWidget {
   final MealPurchaseModel purchaseDetails;
   final MealModel plan;
-  EditMealMenu({this.purchaseDetails, this.plan});
+  PlacedMealMenuOrders({this.purchaseDetails, this.plan});
   @override
-  _EditMealMenuState createState() => _EditMealMenuState();
+  _PlacedMealMenuOrdersState createState() => _PlacedMealMenuOrdersState();
 }
 
-class _EditMealMenuState extends State<EditMealMenu>
+class _PlacedMealMenuOrdersState extends State<PlacedMealMenuOrders>
     with SingleTickerProviderStateMixin {
   final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey<ScaffoldState>();
   ScrollController _scrollController = new ScrollController();
-  TabController _pageController;
-  TextEditingController notes = TextEditingController();
   TextEditingController addressPrimaryLine1 = TextEditingController();
   TextEditingController addressSecondaryLine1 = TextEditingController();
   TextEditingController addressPrimaryLine2 = TextEditingController();
   TextEditingController addressSecondaryLine2 = TextEditingController();
-  final _apiCall = Api.instance;
   FocusNode addressLine1 = FocusNode();
   FocusNode addressLine2 = FocusNode();
-  int menuId = 0;
-  int selectedAddress;
-  double _height = 300;
-  int items = 4;
+  final _apiCall = Api.instance;
+  TabController _pageController;
+  FocusNode addressFocus = FocusNode();
+  List<String> dates = [];
+  List<String> format = [dd, ' ', 'M', ', ', yyyy];
   bool isLoaded = false;
   bool addressSelected = false;
+  String addressType = 'Home';
+  List<String> types = ['Home', 'Work'];
   String addressArea = '';
   String deliveryAddress;
   String localAddress = '';
-  String addressType = 'Home';
-  MenuOrderModel foodItemOrder;
-  List<String> types = ['Home', 'Work'];
+  double _height = 300;
+  DateTime startDate;
+  int items = 4;
+  int menuId = 0;
+  int selectedAddress;
+  List<DateTime> breakDates = List();
+  List<DateTime> planSelectedOffDays = List();
+  List<DateTime> deliverPrimary = List();
+  List<DateTime> deliverSecondary = List();
   List<MenuModel> menuItems = List();
   List<MenuCategoryModel> categoryItems = List();
   List<MenuCategoryModel> mainCategoryItems = List();
   List<MenuCategoryModel> tempCategoryItems = List();
   List<List<MenuCategoryModel>> subCategoryItems = List();
-  List<List<FoodItemModel>> foodItems = List();
-  List<FoodItemModel> itemsFood = List();
-  List<FoodItemModel> expansionFoodItems = List();
-  List<String> dates = [];
+  List<List<MenuOrderModel>> foodItems = List();
+  List<MenuOrderModel> itemsFood = List();
+  List<MenuOrderModel> expansionFoodItems = List();
+
+  @override
+  void initState() {
+    super.initState();
+    menuId = widget.plan.menuId;
+    getDates();
+    getData();
+    _pageController = TabController(
+        length: int.parse(widget.purchaseDetails.mealPlanDuration),
+        vsync: this);
+  }
 
   getData() async {
     await getMenuCategories(menuId);
@@ -72,7 +88,6 @@ class _EditMealMenuState extends State<EditMealMenu>
     });
     categoryItems = await _apiCall.getCategories(menuId);
     for (int i = 0; i < categoryItems.length; i++) {
-      categoryItems[i].showNew();
       if (categoryItems[i].parent == 0) {
         mainCategoryItems.add(categoryItems[i]);
         tempCategoryItems = [];
@@ -103,25 +118,25 @@ class _EditMealMenuState extends State<EditMealMenu>
     if (categoryItems.isNotEmpty) {
       for (int i = 0; i < categoryItems.length;) {
         foodItems.add(await _apiCall
-            .getCategoryFoodItems(
-                menuId.toString(), categoryItems[i].id.toString())
+            .getCurrentMealCategoryOrdersFoodItems(
+                categoryItems[i].id.toString(), widget.purchaseDetails.id)
             .whenComplete(() => i++));
       }
     }
-    print('subCategoryItems.length: ${subCategoryItems.length}');
-    print('mainCategoryItems.length: ${mainCategoryItems.length}');
-    print('categoryItems.length: ${categoryItems.length}');
-    print('foodCategories.length: ${foodItems.length}');
-
-    setState(() {
-      isLoaded = true;
-    });
+    if (mounted) {
+      setState(() {
+        isLoaded = true;
+      });
+    }
   }
 
   getDates() {
     dates = [];
     int count = 0;
-    DateTime startDate = DateTime.parse(widget.purchaseDetails.startDate);
+    startDate = DateTime.parse(widget.purchaseDetails.startDate);
+    DateTime date;
+    startDate = DateTime(
+        startDate.year, startDate.month, startDate.day, 00, 00, 00, 000);
     List days = widget.purchaseDetails.weekdays;
     if (days.contains('Thu')) {
       days.remove('Thu');
@@ -130,54 +145,116 @@ class _EditMealMenuState extends State<EditMealMenu>
     for (int i = 0;
         i < int.parse(widget.purchaseDetails.mealPlanDuration);
         i++) {
-      DateTime date = startDate.add(Duration(days: (count)));
-      while (!days.contains(formatDate(date, [D]))) {
+      date = startDate.add(Duration(days: (count)));
+      while (!days.contains(formatDate(date, [D])) ||
+          ((breakDates != null) ? (breakDates.contains(date)) : false)) {
+        if (!days.contains(formatDate(date, [D]))) {
+          planSelectedOffDays.add(date);
+        }
         count++;
         date = startDate.add(Duration(days: count));
       }
       dates.insert(i, formatDate(date, [dd, '/', mm]));
+      print(date);
       count++;
+    }
+    widget.purchaseDetails.setEndDate(date.toString());
+    print('end Date: ${widget.purchaseDetails.endDate}');
+  }
+
+  void _onSelectionChanged(
+      DateRangePickerSelectionChangedArgs args, bool address) {
+    if (!address) {
+      print(args.value);
+      breakDates = [];
+      breakDates = args.value;
+    } else {
+      if (selectedAddress == 0) {
+        deliverPrimary = args.value;
+        if (args.value != null) {
+          deliverPrimary = [];
+        }
+      } else if (selectedAddress == 1) {
+        deliverSecondary = args.value;
+        if (args.value != null) {
+          deliverSecondary = [];
+        }
+      }
     }
   }
 
-  void initState() {
-    super.initState();
-    menuId = widget.plan.menuId;
-    addressLine1.addListener(() {
-      if (addressLine1.hasFocus) {
-        print('height increased');
-        setState(() {
-          items = 5;
-          _height = 550;
-        });
-      } else if (!addressLine1.hasFocus) {
-        print('height decreased');
-        setState(() {
-          items = 4;
-          _height = 300;
-        });
-      }
+  showCalendar({bool addressCalendar}) async {
+    await showDialog(
+        context: context,
+        barrierDismissible: true,
+        builder: (BuildContext context) {
+          return Center(
+              child: Material(
+            child: Container(
+              height: MediaQuery.of(context).size.height * 0.65,
+              width: MediaQuery.of(context).size.width * 0.85,
+              child: SfDateRangePicker(
+                showNavigationArrow: true,
+                todayHighlightColor: defaultGreen,
+                toggleDaySelection: true,
+                headerHeight: 60,
+                initialSelectedDates: addressCalendar
+                    ? selectedAddress == 0
+                        ? deliverPrimary
+                        : deliverSecondary
+                    : breakDates,
+                selectionColor: defaultGreen,
+                monthCellStyle: DateRangePickerMonthCellStyle(
+                    todayCellDecoration: BoxDecoration(
+                        border: Border.all(color: defaultGreen),
+                        shape: BoxShape.circle),
+                    todayTextStyle: TextStyle(
+                      color: defaultGreen,
+                    ),
+                    specialDatesDecoration: BoxDecoration(
+                        color: Colors.deepPurple,
+                        border: Border.all(color: defaultGreen),
+                        shape: BoxShape.circle),
+                    blackoutDateTextStyle: const TextStyle(
+                        color: Colors.grey,
+                        decoration: TextDecoration.lineThrough)),
+                monthViewSettings: DateRangePickerMonthViewSettings(
+                    numberOfWeeksInView:
+                        ((DateTime.parse(widget.purchaseDetails.endDate))
+                                        .difference(startDate)
+                                        .inDays /
+                                    7)
+                                .round() +
+                            1,
+                    specialDates: addressCalendar
+                        ? selectedAddress == 0
+                            ? deliverPrimary
+                            : deliverSecondary
+                        : breakDates,
+                    blackoutDates: planSelectedOffDays),
+                enablePastDates: false,
+                minDate: DateTime.parse(widget.purchaseDetails.startDate),
+                maxDate: DateTime.parse(widget.purchaseDetails.endDate),
+                view: DateRangePickerView.month,
+                onSelectionChanged: (DateRangePickerSelectionChangedArgs args) {
+                  if (args.value != null) {
+                    _onSelectionChanged(args, addressCalendar);
+                  } else {
+                    setState(() {
+                      breakDates = [];
+                    });
+                  }
+                },
+                selectionMode: DateRangePickerSelectionMode.multiple,
+              ),
+            ),
+          ));
+        }).then((value) {
+      getDates();
+      setState(() {
+        print(breakDates);
+      });
     });
-    addressLine2.addListener(() {
-      if (addressLine2.hasFocus) {
-        print('height increased');
-        setState(() {
-          items = 5;
-          _height = 550;
-        });
-      } else if (!addressLine2.hasFocus) {
-        print('height decreased');
-        setState(() {
-          items = 4;
-          _height = 300;
-        });
-      }
-    });
-    _pageController = TabController(
-        length: int.parse(widget.purchaseDetails.mealPlanDuration),
-        vsync: this);
-    getDates();
-    getData();
   }
 
   void addAddressBottomSheet({int address}) {
@@ -338,8 +415,9 @@ class _EditMealMenuState extends State<EditMealMenu>
                                     ' ' +
                                     addressSecondaryLine2.text;
                               }
+                              Navigator.pop(context);
+                              showCalendar(addressCalendar: true);
                             }
-                            Navigator.pop(context);
                           },
                           child: Container(
                             margin: EdgeInsets.only(top: 20),
@@ -496,175 +574,263 @@ class _EditMealMenuState extends State<EditMealMenu>
               color: defaultGreen,
             ),
           ),
-          title: Text('Meal Plan', style: appBarTextStyle),
+          title: Text('Placed Meal Orders', style: appBarTextStyle),
         ),
         body: Container(
-          child: Column(children: [
-            Expanded(
-              flex: 4,
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  Expanded(
-                    flex: 7,
-                    child: Padding(
-                      padding: const EdgeInsets.only(left: 10.0),
-                      child: Column(
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          Expanded(
-                            flex: 3,
-                            child: Text(
-                              widget.purchaseDetails.mealPlanName,
-                              style: TextStyle(
-                                fontFamily: 'RobotoCondensedReg',
-                                fontSize: 28,
-                                fontWeight: FontWeight.bold,
-                              ),
-                            ),
-                          ),
-                          Expanded(
-                            flex: 7,
-                            child: ClipRect(
-                              child: Text(
-                                widget.plan.details,
-                                style: authInputTextStyle.copyWith(
-                                    fontSize: 14, color: Colors.black),
-                              ),
-                            ),
-                          )
-                        ],
-                      ),
-                    ),
-                  ),
-                  Expanded(
-                      flex: 3,
-                      child: Container(
-                        decoration: BoxDecoration(
-                          color: Colors.white,
-                          shape: BoxShape.circle,
-                          boxShadow: [
-                            BoxShadow(
-                              blurRadius: 4,
-                              color: Colors.grey[500],
-                              spreadRadius: 0,
-                              offset: const Offset(0.0, 0.0),
-                            )
-                          ],
-                        ),
-                        child: CircleAvatar(
-                          radius: 45,
-                          backgroundColor: Colors.white,
-                          child: CachedNetworkImage(
-                            imageUrl: widget.plan.picture ??
-                                "http://via.placeholder.com/350x150",
-                            imageBuilder: (context, imageProvider) => Container(
-                              decoration: BoxDecoration(
-                                shape: BoxShape.circle,
-                                image: DecorationImage(
-                                  image: imageProvider,
-                                  fit: BoxFit.cover,
+          padding: EdgeInsets.only(top: 10),
+          child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              mainAxisAlignment: MainAxisAlignment.start,
+              children: [
+                Expanded(
+                  flex: 5,
+                  child: Padding(
+                    padding: const EdgeInsets.symmetric(horizontal: 10.0),
+                    child: Column(
+                      children: [
+                        Flexible(
+                          fit: FlexFit.loose,
+                          flex: 2,
+                          child: Row(
+                            mainAxisAlignment: MainAxisAlignment.center,
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              Flexible(
+                                flex: 7,
+                                child: Column(
+                                  mainAxisAlignment: MainAxisAlignment.start,
+                                  crossAxisAlignment: CrossAxisAlignment.start,
+                                  children: [
+                                    Flexible(
+                                      child: Text(
+                                        widget.purchaseDetails.mealPlanName,
+                                        style: TextStyle(
+                                          fontFamily: 'RobotoCondensedReg',
+                                          fontSize: 28,
+                                          fontWeight: FontWeight.bold,
+                                        ),
+                                      ),
+                                    ),
+                                    Flexible(
+                                      child: Padding(
+                                        padding: EdgeInsets.symmetric(
+                                            horizontal: 10),
+                                        child: Text(
+                                          widget.plan.details,
+                                          style: authInputTextStyle.copyWith(
+                                              fontSize: 14,
+                                              color: Colors.black),
+                                        ),
+                                      ),
+                                    ),
+                                  ],
                                 ),
                               ),
-                            ),
-                            placeholder: (context, url) =>
-                                CircularProgressIndicator(),
-                            errorWidget: (context, url, error) =>
-                                Icon(Icons.error),
+                              Flexible(
+                                  flex: 3,
+                                  child: Container(
+                                    decoration: BoxDecoration(
+                                      color: Colors.white,
+                                      shape: BoxShape.circle,
+                                      boxShadow: [
+                                        BoxShadow(
+                                          blurRadius: 4,
+                                          color: Colors.grey[500],
+                                          spreadRadius: 0,
+                                          offset: const Offset(0.0, 0.0),
+                                        )
+                                      ],
+                                    ),
+                                    child: CircleAvatar(
+                                      radius: 45,
+                                      backgroundColor: Colors.white,
+                                      child: CachedNetworkImage(
+                                        imageUrl: widget.plan.picture ??
+                                            "http://via.placeholder.com/350x150",
+                                        imageBuilder:
+                                            (context, imageProvider) =>
+                                                Container(
+                                          decoration: BoxDecoration(
+                                            shape: BoxShape.rectangle,
+                                            image: DecorationImage(
+                                              image: imageProvider,
+                                              fit: BoxFit.cover,
+                                            ),
+                                          ),
+                                        ),
+                                        placeholder: (context, url) =>
+                                            CircularProgressIndicator(),
+                                        errorWidget: (context, url, error) =>
+                                            Icon(Icons.error),
+                                      ),
+                                    ),
+                                  ))
+                            ],
                           ),
                         ),
-                      ))
-                ],
-              ),
-            ),
-            Expanded(
-                flex: 1,
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                  children: [
-                    Text('${widget.purchaseDetails.kCal} Calorie'),
-                    TextButton(
-                        onPressed: () {
-                          getBottomSheet();
-                        },
-                        child: Text(
-                          'Address',
-                          style: TextStyle(
-                            color: darkGreen,
-                            decoration: TextDecoration.underline,
-                          ),
-                        )),
-                  ],
-                )),
-            Expanded(
-              flex: 1,
-              child: Container(
-                decoration: BoxDecoration(
-                  color: white,
-                  boxShadow: [
-                    BoxShadow(
-                      blurRadius: 4,
-                      color: Colors.black.withOpacity(0.25),
-                      spreadRadius: 0,
-                      offset: const Offset(0.0, 0.0),
-                    )
-                  ],
+                      ],
+                    ),
+                  ),
                 ),
-                child: TabBar(
-                  isScrollable: true,
-                  controller: _pageController,
-                  onTap: (index) {},
-                  labelStyle:
-                      selectedTab.copyWith(color: Colors.black, fontSize: 14),
-                  indicatorColor: Colors.transparent,
-                  indicatorWeight: 3.0,
-                  indicatorSize: TabBarIndicatorSize.label,
-                  labelColor: Colors.black,
-                  labelPadding: EdgeInsets.symmetric(horizontal: 10),
-                  unselectedLabelStyle:
-                      unSelectedTab.copyWith(color: Colors.grey),
-                  unselectedLabelColor: Colors.grey,
-                  tabs: List.generate(
-                      int.parse(widget.purchaseDetails.mealPlanDuration),
-                      (index) {
-                    return Tab(
+                Expanded(
+                    flex: 3,
+                    child: Padding(
+                      padding: const EdgeInsets.only(left: 30.0),
                       child: Column(
-                        crossAxisAlignment: CrossAxisAlignment.center,
-                        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                        mainAxisSize: MainAxisSize.min,
+                        crossAxisAlignment: CrossAxisAlignment.start,
+                        mainAxisAlignment: MainAxisAlignment.start,
                         children: [
                           Flexible(
-                              fit: FlexFit.loose,
-                              child: Text('Day ${(index + 1).toString()}')),
+                            child: Text(
+                              "${widget.purchaseDetails.mealPlanDuration} day meal plan • " +
+                                  (widget.plan.type == 0
+                                      ? "With Weekends"
+                                      : "Without Weekends"),
+                              style: TextStyle(color: Colors.black),
+                            ),
+                          ),
                           Flexible(
-                              fit: FlexFit.loose, child: Text(dates[index]))
+                            fit: FlexFit.loose,
+                            child: Text(
+                              "${widget.purchaseDetails.kCal} Calorie",
+                              style: TextStyle(color: Colors.black),
+                            ),
+                          ),
+                          Flexible(
+                            child: Text(
+                              'Start Date - ${formatDate(DateTime.parse(widget.purchaseDetails.startDate), format)}',
+                              style: TextStyle(color: Colors.black),
+                            ),
+                          ),
                         ],
                       ),
-                    );
-                  }),
+                    )),
+                Expanded(
+                  flex: 2,
+                  child: Row(
+                    mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                    crossAxisAlignment: CrossAxisAlignment.center,
+                    children: [
+                      TextButton(
+                          onPressed: () async {
+                            getBottomSheet();
+                          },
+                          child: Text(
+                            'Address',
+                            style: TextStyle(
+                              color: darkGreen,
+                              decoration: TextDecoration.underline,
+                            ),
+                          )),
+                      Text('${widget.purchaseDetails.kCal} Calorie'),
+                      TextButton(
+                          onPressed: () async {
+                            await showCalendar(addressCalendar: false);
+                          },
+                          child: Text(
+                            'Breaks',
+                            style: TextStyle(
+                              color: darkGreen,
+                              decoration: TextDecoration.underline,
+                            ),
+                          )),
+                    ],
+                  ),
                 ),
-              ),
-            ),
-            Expanded(
-              flex: 9,
-              child: isLoaded
-                  ? TabBarView(
+                Expanded(
+                  flex: 2,
+                  child: Container(
+                    decoration: BoxDecoration(
+                      color: white,
+                      boxShadow: [
+                        BoxShadow(
+                          blurRadius: 4,
+                          color: Colors.black.withOpacity(0.25),
+                          spreadRadius: 0,
+                          offset: const Offset(0.0, 0.0),
+                        )
+                      ],
+                    ),
+                    child: TabBar(
+                      isScrollable: true,
                       controller: _pageController,
-                      children: List.generate(
+                      onTap: (index) {},
+                      labelStyle: selectedTab.copyWith(
+                          color: Colors.black, fontSize: 14),
+                      indicatorColor: Colors.transparent,
+                      indicatorWeight: 3.0,
+                      indicatorSize: TabBarIndicatorSize.label,
+                      labelColor: Colors.black,
+                      labelPadding: EdgeInsets.symmetric(horizontal: 10),
+                      unselectedLabelStyle:
+                          unSelectedTab.copyWith(color: Colors.grey),
+                      unselectedLabelColor: Colors.grey,
+                      tabs: List.generate(
                           int.parse(widget.purchaseDetails.mealPlanDuration),
                           (index) {
-                        return Container(
-                          child: menuUi(day: index + 1),
+                        return Tab(
+                          child: Column(
+                            crossAxisAlignment: CrossAxisAlignment.center,
+                            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                            mainAxisSize: MainAxisSize.min,
+                            children: [
+                              Text('Day ${(index + 1).toString()}'),
+                              Text(dates[index])
+                            ],
+                          ),
                         );
-                      }))
-                  : Center(child: CircularProgressIndicator()),
-            )
-          ]),
+                      }),
+                    ),
+                  ),
+                ),
+                Expanded(
+                  flex: 18,
+                  child: isLoaded
+                      ? TabBarView(
+                          controller: _pageController,
+                          children: List.generate(
+                              int.parse(
+                                  widget.purchaseDetails.mealPlanDuration),
+                              (index) {
+                            return Column(
+                              children: [
+                                Expanded(
+                                  child: Container(
+                                    child: menuUi(day: index + 1),
+                                  ),
+                                ),
+                                GestureDetector(
+                                  onTap: () {
+                                    Navigator.push(
+                                        context,
+                                        CupertinoPageRoute(
+                                            builder: (BuildContext context) =>
+                                                PlaceMealMenuOrders(
+                                                    plan: widget.plan,
+                                                    purchaseDetails: widget
+                                                        .purchaseDetails)));
+                                  },
+                                  child: Container(
+                                    height: 50,
+                                    margin: EdgeInsets.only(top: 20),
+                                    color: defaultGreen,
+                                    child: Center(
+                                        child: Text(
+                                      'Edit Menu',
+                                      style: selectedTab.copyWith(color: white),
+                                    )),
+                                  ),
+                                ),
+                              ],
+                            );
+                          }))
+                      : Center(child: CircularProgressIndicator()),
+                )
+              ]),
         ));
   }
 
-  Widget foodItemCard(FoodItemModel item) {
+  Widget foodItemCard(MenuOrderModel item) {
     return Container(
       height: 123,
       margin: EdgeInsets.symmetric(horizontal: 10),
@@ -678,19 +844,8 @@ class _EditMealMenuState extends State<EditMealMenu>
                 mainAxisAlignment: MainAxisAlignment.center,
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
-                  Container(
-                    height: 15,
-                    width: 15,
-                    child: Image.asset(
-                      item.isVeg ? 'images/veg.png' : 'images/nonVeg.png',
-                      fit: BoxFit.fitHeight,
-                    ),
-                  ),
-                  SizedBox(
-                    height: 10,
-                  ),
                   Text(
-                    item.foodName ?? "Food Name",
+                    item.foodItemName ?? "Food Name",
                     style: appBarTextStyle.copyWith(fontSize: 20),
                   ),
                   SizedBox(
@@ -719,8 +874,7 @@ class _EditMealMenuState extends State<EditMealMenu>
                       width: 82,
                       decoration: authFieldDecoration,
                       child: CachedNetworkImage(
-                        imageUrl: item.picture ??
-                            "http://via.placeholder.com/350x150",
+                        imageUrl: "http://via.placeholder.com/350x150",
                         imageBuilder: (context, imageProvider) => Container(
                           height: 60,
                           decoration: BoxDecoration(
@@ -737,70 +891,6 @@ class _EditMealMenuState extends State<EditMealMenu>
                         ),
                       ),
                     )),
-                Positioned(
-                  top: 70.0,
-                  left: 3.5,
-                  child: SizedBox(
-                    width: 75,
-                    height: 33,
-                    child: Container(
-                      decoration: BoxDecoration(
-                        boxShadow: [
-                          BoxShadow(
-                            blurRadius: 4,
-                            color: Colors.black.withOpacity(0.25),
-                            spreadRadius: 0,
-                            offset: const Offset(0.0, 0.0),
-                          )
-                        ],
-                      ),
-                      child: TextButton(
-                        onPressed: () async {
-                          if (!item.isSelected) {
-                            if (addressSelected) {
-                              foodItemOrder = MenuOrderModel(
-                                  foodItemId: item.id,
-                                  foodItemCategoryId: item.categoryId,
-                                  mealPurchaseId:
-                                      int.parse(widget.purchaseDetails.id),
-                                  menuItemDate: item.date,
-                                  menuItemDay: item.day,
-                                  foodItemName: item.foodName,
-                                  deliveryAddress: deliveryAddress,
-                                  note:
-                                      notes.text.isEmpty ? "null" : notes.text);
-                              String body =
-                                  convert.jsonEncode(foodItemOrder.toMap());
-                              print(body);
-                              int result =
-                                  await _apiCall.postMenuOrder(foodItemOrder);
-                              if (result != null) {
-                                item.updateOrderItemId(result);
-                                item.change(true);
-                                setState(() {});
-                              }
-                            } else {
-                              _scaffoldKey.currentState.showSnackBar(SnackBar(
-                                  content: Text('Select address first')));
-                            }
-                          } else {
-                            // TODO: delete selected item
-                            item.change(false);
-                            setState(() {});
-                          }
-                        },
-                        child: Text(item.isSelected ? 'Selected' : 'Select',
-                            style: selectedTab.copyWith(
-                                color: item.isSelected ? white : defaultGreen,
-                                fontSize: 16)),
-                        style: TextButton.styleFrom(
-                          backgroundColor:
-                              item.isSelected ? defaultGreen : white,
-                        ),
-                      ),
-                    ),
-                  ),
-                )
               ],
             ),
           )
@@ -823,47 +913,8 @@ class _EditMealMenuState extends State<EditMealMenu>
             ),
             initiallyExpanded: indexMajor == 0 ? true : false,
             children: subCategoryItems[indexMajor].length == 0
-                ? List.generate(itemsFood.length + 1, (index) {
-                    if (index == itemsFood.length) {
-                      return Column(
-                        children: [
-                          Divider(),
-                          Padding(
-                            padding:
-                                const EdgeInsets.symmetric(horizontal: 15.0),
-                            child: Text(
-                              'Let us know if there is something you’d want us to know about your menu, we’ll pass it on to the chef.',
-                              style: authInputTextStyle.copyWith(
-                                  fontSize: 20, color: Colors.black),
-                            ),
-                          ),
-                          Container(
-                              margin: EdgeInsets.all(15),
-                              padding: EdgeInsets.symmetric(horizontal: 10),
-                              height: MediaQuery.of(context).size.height * 0.15,
-                              decoration: BoxDecoration(
-                                color: white,
-                                boxShadow: [
-                                  BoxShadow(
-                                    blurRadius: 4,
-                                    color: Colors.black.withOpacity(0.25),
-                                    spreadRadius: 0,
-                                    offset: const Offset(0.0, 0.0),
-                                  )
-                                ],
-                              ),
-                              child: TextFormField(
-                                controller: notes,
-                                onChanged: (value) {
-                                  notes.text = value;
-                                },
-                                decoration: authInputFieldDecoration.copyWith(
-                                    hintText: 'Enter your note here'),
-                              )),
-                        ],
-                      );
-                    }
-                    if (itemsFood[index].day == day) {
+                ? List.generate(itemsFood.length, (index) {
+                    if (itemsFood[index].menuItemDay == day) {
                       return foodItemCard(itemsFood[index]);
                     }
                     return Container();
@@ -885,50 +936,8 @@ class _EditMealMenuState extends State<EditMealMenu>
                           style: selectedTab.copyWith(fontSize: 20),
                         ),
                         initiallyExpanded: indexMinor == 0 ? true : false,
-                        children: List.generate(itemsFood.length + 1, (index) {
-                          if (index == itemsFood.length) {
-                            return Column(
-                              children: [
-                                Divider(),
-                                Padding(
-                                  padding: const EdgeInsets.symmetric(
-                                      horizontal: 15.0),
-                                  child: Text(
-                                    'Let us know if there is something you’d want us to know about your menu, we’ll pass it on to the chef.',
-                                    style: authInputTextStyle.copyWith(
-                                        fontSize: 20, color: Colors.black),
-                                  ),
-                                ),
-                                Container(
-                                    margin: EdgeInsets.all(15),
-                                    padding:
-                                        EdgeInsets.symmetric(horizontal: 10),
-                                    height: MediaQuery.of(context).size.height *
-                                        0.15,
-                                    decoration: BoxDecoration(
-                                      color: white,
-                                      boxShadow: [
-                                        BoxShadow(
-                                          blurRadius: 4,
-                                          color: Colors.black.withOpacity(0.25),
-                                          spreadRadius: 0,
-                                          offset: const Offset(0.0, 0.0),
-                                        )
-                                      ],
-                                    ),
-                                    child: TextFormField(
-                                      controller: notes,
-                                      onChanged: (value) {
-                                        notes.text = value;
-                                      },
-                                      decoration:
-                                          authInputFieldDecoration.copyWith(
-                                              hintText: 'Enter your note here'),
-                                    )),
-                              ],
-                            );
-                          }
-                          if (itemsFood[index].day == day) {
+                        children: List.generate(itemsFood.length, (index) {
+                          if (itemsFood[index].menuItemDay == day) {
                             return foodItemCard(itemsFood[index]);
                           }
                           return Container();
@@ -941,6 +950,23 @@ class _EditMealMenuState extends State<EditMealMenu>
           color: Colors.black12,
         );
       },
+    );
+  }
+}
+
+class CustomWeekdayLabelsRow extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: <Widget>[
+        Expanded(child: Text("M", textAlign: TextAlign.center)),
+        Expanded(child: Text("T", textAlign: TextAlign.center)),
+        Expanded(child: Text("W", textAlign: TextAlign.center)),
+        Expanded(child: Text("T", textAlign: TextAlign.center)),
+        Expanded(child: Text("F", textAlign: TextAlign.center)),
+        Expanded(child: Text("S", textAlign: TextAlign.center)),
+        Expanded(child: Text("S", textAlign: TextAlign.center)),
+      ],
     );
   }
 }
