+++ pubspec.yaml
@@ -21,6 +21,7 @@ dependencies:
   flutter_dropdown: ^0.0.8
   date_format: ^1.0.9
   cached_network_image: ^2.5.0
+  syncfusion_flutter_datepicker: ^18.4.41-beta
 
 dev_dependencies:
   flutter_test:
